Main: VC++ 19.2 update 11 build adjustments (lossy floating point conversions). bugref:8489
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@93309 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/GuestSessionImplTasks.cpp
@@ -443,7 +443,7 @@ int GuestSessionTask::fileCopyFromGuestInner(const Utf8Str &strSrcFile, ComObjPt
             && fCanceled)
             break;
 
-        rc = setProgress((ULONG)(cbWrittenTotal / ((uint64_t)cbSize / 100.0)));
+        rc = setProgress((ULONG)((double)cbWrittenTotal / (double)cbSize / 100.0));
         if (RT_FAILURE(rc))
             break;
     }
@@ -763,7 +763,7 @@ int GuestSessionTask::fileCopyToGuestInner(const Utf8Str &strSrcFile, RTVFSFILE 
             && fCanceled)
             break;
 
-        rc = setProgress((ULONG)(cbWrittenTotal / ((uint64_t)cbSize / 100.0)));
+        rc = setProgress((ULONG)((double)cbWrittenTotal / (double)cbSize / 100.0));
         if (RT_FAILURE(rc))
             break;
     }
